The basic mechanisms of Tetris are implemented
However the failure detection is not always working.

Includes piece prediction, i.e. where the piece would fall if it is
dropped, indicated by a transparent piece. Also displays the next piece
that will be given to the player.

To do: display score, introduce concept of levels, allow the player to
store a piece and properly detect when a player has lost.